requirements.txt
camelot_py==1.0.0
pandas==2.3.0
plotly==6.1.2
PyYAML==6.0.2
PyYAML==6.0.2
streamlit==1.45.1
